Streamlined all datascrapers
