+++ apis/starsavn/classes/extras.py
@@ -1,6 +1,7 @@
 import copy
 from itertools import chain
 import math
+from pathlib import Path
 from typing import Any, Literal, Optional, Union
 
 
@@ -175,12 +176,20 @@ class endpoint_links(object):
         return final_links
 
 
-# Lol?
 class ErrorDetails:
-    def __init__(self, result) -> None:
+    def __init__(self, result: dict[str, Any]) -> None:
         error = result["error"] if "error" in result else result
         self.code = error["code"]
         self.message = error["message"]
+
+    async def format(self, extras: dict[str, Any]):
+        match self.code:
+            case 0:
+                match self.message:
+                    case "User not found":
+                        link = Path(extras["link"])
+                        self.message = f"{link.name} not found"
+        return self
 
 
 def create_headers(
