+++ apis/starsavn/classes/message_model.py
@@ -60,7 +60,7 @@ class create_message:
         )
         return result
 
-    async def link_picker(self, media, video_quality):
+    async def link_picker(self, media: dict[str, Any], video_quality: str):
         link = ""
         if "source" in media:
             quality_key = "source"
@@ -77,9 +77,6 @@ class create_message:
                             if quality_link:
                                 link = quality_link
                                 break
-                            print
-                        print
-                    print
         if "src" in media:
             link = media["src"]["source"]
         return link
