+++ apis/starsavn/classes/story_model.py
@@ -22,7 +22,7 @@ class create_story:
         self.placedContents: list = option.get("placedContents")
         self.answered: int = option.get("answered")
 
-    async def link_picker(self, media, video_quality):
+    async def link_picker(self, media: dict[str, Any], video_quality: str):
         link = ""
         if "source" in media:
             quality_key = "source"
@@ -39,9 +39,6 @@ class create_story:
                             if quality_link:
                                 link = quality_link
                                 break
-                            print
-                        print
-                    print
         if "src" in media:
-            link = media["src"]
+            link = media["src"]["source"]
         return link
